Removed JS dashboard thumbnail generation & seperated concerns in api and services
