+++ viewer/src/api/project.js
@@ -6,14 +6,3 @@ export const fetchProject = async () => {
         return null
     }
 }
-
-export const fetchExplorer = async () => {
-    const response = await fetch('/data/explorer.json');
-    if (response.status === 200) {
-        const data = await response.json();
-        return data;
-    } else {
-        console.error('Failed to fetch explorer data');
-        return null
-    }
-}
