+++ viewer/src/services/thumbnailService.js
@@ -1,0 +1,35 @@
+const sanitizeDashboardName = (name) => {
+    return name.replace(/[^a-zA-Z0-9]/g, '_');
+};
+
+export const fetchDashboardThumbnail = async (dashboardName) => {
+    try {
+        const safeName = sanitizeDashboardName(dashboardName);
+        const response = await fetch(`/api/thumbnails/${safeName}`);
+        
+        if (!response.ok) {
+            throw new Error('Thumbnail not found');
+        }
+        
+        const data = await response.json();
+        return data.thumbnail;
+    } catch (error) {
+        // Silently fail for missing thumbnails
+        return null;
+    }
+};
+
+export const fetchDashboardThumbnails = (dashboards, onThumbnailLoaded) => {
+    if (!dashboards || dashboards.length === 0) {
+        return;
+    }
+
+    dashboards.forEach(dashboard => {
+        fetchDashboardThumbnail(dashboard.name)
+            .then(thumbnail => {
+                if (thumbnail) {
+                    onThumbnailLoaded(dashboard.name, thumbnail);
+                }
+            });
+    });
+}
